Update eventDefinition to support dynamic attributes
Adds an 'Extra' field and json Marshaler,Unmarshaler functions
to allow an event to contain any number of properties.  If the property
is not one of the required properties, it is stored in Event.Extra[] when
Unmarshaled from JSON. when Marshaling back to json, the Extra properties
are returned to be top level properties.