refactor code to simplify logic
- Major refactor of the AggMetric struct. Remove FirstTs, LastTs, FirstT0,
LastT0 fields and replace with a single CurrentChunkPos to keep track of
where in the circular Buffer we are currently writing to.  With this new
approach the logic in the Get method was able to be drastically simplified.

- moved the logic for skipping metric that are older then what has already
been written from AggMetric and into Chunk as it is up to the Chunk to
decide what can and can't be written.

- fixed up unit tests which were failing due to the statsd metrics not
being initialized

- polished up the saving/loading of state to disk.