fix to support future data points
We decided we want to support data in the future (forecasts etc).
This means these values need updating.
in the last case it's not a necessity but just looks nicer.

Backstory:

the old code had an optimisation so that in this example case..

numChunks is 10, chunkSpan is 2h,
server started at 00:00, oldestChunk.T0 is 00:200
we only have 4 chunks
you do a query for 1AM to 2AM.
-> for some reason no data was sent between midnight and 2AM.

.. it would return an oldest value of 00:00 instead of 2AM, because
in such cases we could confidently say we would have had the data, if
the data had been sent, and avoid needless cassandra queries.

However we realized this doesn't work once you support sending data in
the future. (i.e. at 11:55 you send data for 1AM), the server wouldn't
have gotten it.  And we also realized we wanted to support sending
future data, this seems useful for forecasts etc.

This optimization got removed in
98b0bf1b0cb70aa819a0a252c51222ca01c2f95b