bugfix: don't create new chunks that have lower t0 than last
since 98b0bf1b0cb70aa819a0a252c51222ca01c2f95b, we moved the
"point must be newer than lastTs" logic into the chunk.
But there's still the case where a point is so much older than the last
one, that it maps to an older t0.
In this case we should not recreate a chunk for the given t0, since we
most likely had previously created, finished and stored it already, and
moved on.