Ensure that previous chunks are persited before trying new chunks.
issue #65
When persiting a chunk, we now first check that the previous chunk
has been persited. If it hasn't been saved we block until it has
been.  The check cascades through all chunks ensuring that all
older chunks are saved before newer ones.

To maintain performance during normal operation, when only the
current Chunk needs persiting it is saved asynchronously and any
failures are ignored.  If the current chunk fails to save, it will
be retried when the next chunk tries to save, but on the second
attempt it will block until the save is complete.

This blocking approach ensures that if there is a short backend
storage outage, performance wont be impacted. But if the backend
outage lasts longer then *ChunkSpan* then processing of new
messages will block, allowing them to spool on the message bus
which has persistant storage.