refactor blocking code.
New design uses a buffered channel as a per series write queue.
A single goroutine reads from the channel and processes the writes
in the order they are written to the channel.  The reader only moves
onto the next write once the first completes successfully.  The user
can set the the size of the writeQueue as a config option.  The
writeQueue can be any size including greater then the numChunks setting.
Once the queue is full, calls to persist() will block.