add http handler for "/"
This handler is designed to be used by LB healthchecks. The handler
will return a 503 unavailable response if the node is not the
primary node and has not been running for at least warmUpPeriod
seconds.  This will ensure that metric Get requests will only be
sent to the node when it has sufficient data to serve.