persist old chunks that have not been saved.
When there is no primary node (due to crash or shutdown) chunks
will not be getting saved. When a secondary node gets promoted
to primary there may be older chunks that never got saved and this
commit ensures that we check for this condition.  If there are older
chunks that need saving, we add them to the writeQueue. Older chunks
are added to the writeQueue before newer chunks.