add warmup.
If the current index is empty or stale, then we will overload
ES if we try and index all metrics when the first arrive.  This
commit adds a warmup period to allow the index to be populated,
without overloading ES.  If the index is empty, it will take up
to 30minutes for all metrics to be available for graphite querying.