only persist chunks asynchronously
i suspect that calling persist() when we step into a new chunkspan
is an operation that:
1) is called for a whole lot of metrics at nearly the same time
2) creates a lot of lock contention by simultaneously trying to add
   to the store/write queues

I don't have hard facts to prove this yet.
But this approach removes this assumed burden and instead relies
solely on the GC tasks to persist chunks.
At least the GC does this metric by metric.
You may want to increase your GC frequency to ~ chunkSpan

we loose the NSQ backpressure on full-queue, perhaps it could be
re-added.  With this approach it becomes more a matter of verifying
GC can keep up, which will need a bit more instrumentation.

to do this we need some refactoring for GC & persister:

* it's now GC's job to not only persist the latest chunk when it
  hasn't been written to in a while, it also has to make sure that
  any previous chunks are being persisted.
* since we now run persist solely as part of a maintenance task, and no longer
  at the exact point where we know we need to persist a metric,
  it is now being invoked with different semantics: the specified pos
  refers to a chunk we may want to save, subject to the same constraints
  as the others. it is not guaranteed unlike before.
* bonus: don't do needless work when we're not primary, in particular
  we only need to finish a chunk when it's being persisted