bugfix: prevent clobbering the output of DecodeMetricData
All concurrent handlers were using a shared pool of MetricDataArray's,
which were used to decode the messagepack data.
When constructing the Metrics array for MetricData, we were using
a slice that had the same backing array as used in the temporary
MetricDataArray's.

This means that inside HandleMessage, when DecodeMetricData returned,
and we wanted to use the MetricData object to add it to defcache,
AggMetrics, etc, the actual metrics could be accidentally overwritten
by other handlers invoking decoders that were using MetricDataArray's
backed by the same underlying array.

The arrays are safe for reuse within each handler from one invocation of
DecodeMetricData to the next, but not across multiple concurrent
handlers.

So a sync.Pool for each handler would do it, but in fact we only need 1
temporary array per handler, and in fact it turns out we already have
one (in the output struct), so now we just use that.