don't keep reallocating MetricData's
instead, a handler instance can just have one MetricData in which
we keep decoding the metrics.
note that this means we must run separate handler instances,
if we would run the same handler concurrently it would be racey
and incorrect. this is a small price to pay as handlers are very
lightweight.

this reduces the rate at which we allocate data (in an ingest-only
workload) by 30%, reduces GC runs slightly,
as well as heap and cpu usage.
see charts @ https://github.com/raintank/raintank-metric/pull/195