merge series when multiple metricDefs for a single target
Metrics can change over time while still maintaining the same
metric.Name. This can happen when tags are changed or when the
interval of a metric changes.  When this happens we need to merge
the multiple sets of metric points into a single set.

Though there can be more then one metricDef for a series, only
one of the metricDefs should be receiving data at any point in time.
So the approach for merging is to use the first non-null value
found for each timestamp in the result.