preriodically update metricDefs in cassandra.
As we now expire stale metrics, regularily update the metricDefs in
Cassandra, so on reload we have an accurate age for them. To ensure
that we dont try and update all metricDefs at the same time we use
a updateFuzzyness value to randomly distribute updates over time.
eg. using an update-interval of 2hours and a updateFuzzyness of 0.5
metricDefs will be updated between every 2 to 3hours.