filter search results by from TS
- update the metricIdx.Find() method to take a third paramater
that is a timestmap that the series should have been updated since.
- If a series has not been seen for a long time, then it will be
excluded from search results. This is separate to the index pruning
that deletes series from the index.
- if a series has multiple metricDefs, ie due to a interval change,
this new feature will ensure that only the relevent metricDef is used
and the old stale one is excluded from the search result. If a the query
is for a time range from when the old metricDef was still being recieved,
then it will also be included.