fix #323
when we quantize, sometimes we need an earlier point to move to the
right (into the requested range), but the point we need fell before
the requested range.
So we just fetch `interval-1` more when needed to make sure we have the
right point in case this happens.
We can just extend the use of the aggSpan (now interval) argument of
getSeries to always refer to the interval of the requested archive
(whether aggregated or not)