Remove unnecessary error
If busybox is not in use the wait_for_endpoint.sh script shows an error
because basename is used on an empty string. This makes the busybox
checking a little more complex, but it gets rid of that error because it
avoids that case where basename is called with an empty string.