fix #323 better
with this approach we make more precise adjustments to from/to
which has two advantages:

1) we now query a slightly narrower timerange, which can be faster
2) historically, fix() has always been needed after getSeries().
   the previous patch relied on this and also requires it in order to:
   A provide the correct amount of points to the user
     (fix removes unneeded points out of the range)
   B provide the user points with clean, quantized timestamps that
     actually lie in the requested range

   With the new patch, we only require calling fix() after getSeries()
   for B, not for A.  So it's still a requirement, and we still have to
   honor it.

Technically, if you only wanted to query for points between a given
range and did not care about B, we don't need to call fix to
"pretty up" the output, and could bypass fix() and be more efficient.
Maybe later we can make it optional.