adopt nsq bench to in bench, remove bad test (don't port to in)
==== the unit test ====
this test isn't very good.
We're testing effectively In and idx, not really nsq (we just mimic it)
So we could change it to call in.process() or in.Handle() concurrently
and test that.

(intermezzo: then we should do all the allocations, encodings, etc
up front, and move out the lock-protected idToId stuff, to
minimize dragging on the duration of the concurrent code, which
maximizes our chances of finding race conditions.)

But, we only really verify that will be get back the right metricdef for
a given id, but we never inserted different metricdefs for a given id.
We could do that, even though metricdefinitions don't really change,
for a given id (id is the hash from all properties), so it wouldn't
be very realstic, but it could still be worthwhile to test the code
this way, since we do plan to add mutable properties later.

However, we would then need to track (as we currently do) how the
last metricdef we inserted for a given id, looks like.  Which requires
locking.  Which we currently have, except the call to process() should
also fall under this lock, otherwise the goroutine could be interrupted
in our tracking information could not match when In actually received
the data.  But adding process() to the critical section of the lock
defeats of course the whole point, as that will make it impossible
to find any race related bugs.

TLDR:
I think this test was flawed to begin with: we're hoping to expose
race conditions which requires concurrent random input, but we need
to track the order of the input, which requires locking around the
thing we're testing.  I guess the only way to expose these issues
is with running it with -race enabled, and just reading the code
carefully :]

==== the benchmark ====
can be repurposed as a benchmark for in.Handle, which is now the most
common approach (e.g. used by kafka-mdm)

Handle is mostly just the sum of its parts, e.g.:
* metric validation
* statsd calls
* index Add calls
* adding data to AggMetrics and AggMetric
* adding to usage

which we already benchmark anyway, but it doesn't hurt to bench them
together.