always run benchmarks run on circleCI push
they take 10 minutes and they are not all accurate.
we can make this better as we go along.

it sucks when benchmarks break when we don't test them
also, this can be the first step towards tracking performance numbers
over time. maybe.. the cloud probably doesn't deliver consistent enough
performance, but we can see.