only add metricdef to DefsById if we can also add it to the tree
otherwise we accidentally add defs to DefsById and Add() will
always think it's already in the tree even when it's not.
So when you then remove the cause of a metricdef not being able
to go into the tree as in #397, it would no longer add it.

This also changes the behavior of the idx.memory.fail metric:
it will now keep triggering for every attempt to add to the
index which fails.  This is more accurate.

fix #397 - fix #399