Refactor iterators to add a way to get size of data
Previously our iter.Iter structs only referred to iterators returned by
the go-tsz packages. The problem we're trying to solve is that from the
go-tsz iterators there's no way to get the length of the underlying
data, but we need to know that to limit the size of the chunk cache.
This patch passes a length attribute to each instance of iter.Iter,
along with the go-tsz iterator.

This means we can get the length of the underlying byte slice from an
iterator, which will later be necessary for the chunk cache (#401).

The commit also includes some changes done by gofmt.