s/endTs/nextTs/
this is a bit more clear that "endTs" is really the ts that is the
beginning of the next chunk.
If we used continuous math (and/or if everyone thought of our timestamps
as continuous) then we could have used the same name.

However since our timestamps are discrete and at second resolution,
it's more clear this way. (the end of a chunk is one second prior to the
t0 of the next)