figure out previous chunk on cache add
in a scenario where a metric keeps getting queried for the last few
seconds (less than one chunk span) this chunk would get added into the
cache without the previous chunk being known. with this change the
`Add()` method uses the already existing `seekDesc()` method to check if
the previous chunk is in the cache and if so they get connected
accordingly.

the `searchBackward()` method gets changed to not search further than
the given `from`.