I use hashicorp/memberlist for peer discovery and management
memberlist is a library that manages cluster membership and
member failure detection using a gossip based protocol.

To enable use of memberlist a "ClusterManager" struct has been created.
This struct handles co-ordination of changes to the state of every node.

Nodes are now passed throughout the codebase by value (copies) which eliminates
the need for complex locking.  To prevent races when reading/updating the localNode
properties, all changes are co-ordinated through the ClusterManager which provides
the required locking.