reset cache instead of reinstantiating it in tests
previously some of the cache tests reinstantiated the chunk cache and
it's accounting many times per run. this consumed large amounts of
memory, especially with the `-race` flag, and made circle ci hit the 4G
limit.
this change adds a reset method and uses it in the unit test, which
leads to a drastic reduction in memory usage in the test.