set mode by string
when using a bool to set a node to primary/secondary we run into the
problem that the `0` value of bool is `false` so when specifying
`false` the parsing library errors saying that a required parameter
has not been passed.
now we're taking the value as a string instead and then parse it into
a bool.

fixes #431