Don't add metrics data if they can't be added to the index
Note this requires an API change:
If we restrict our index implementations to ones that are built around the memory index (which seems to be the case for the foreseeable future)
then we don't do Get operations that can fail (result in an error condition) due to io operations failing or something.
So Get cannot run into error conditions, so we can simplify the Get signature to return a 'found' bool instead of an error.
Since AddOrUpdate first calls Get, and then tries to add to the index,
this lets us call AddOrUpdate, knowing that errors will only happen if something's wrong with the metric and it can't be added to the index,
and no longer due to the Get() step.
Which in turn means we can confidently say there's no point adding the data in that case, as the error cannot be due to a transient condition,
but rather something fundamentally wrong why we can't accept it.

Once the metric can be added to the index, we save the data.