fix garbled output
while executing a template concurrently is OK, and writes to os.Stdout
are atomic as well, we had a bug:
Template.Execute does separate writes for each "step", i.e. each
variable that needs to be printed, and the final newline.
So even under a medium workload of 10kMsg/s occasionally the output
would be garbled due to attributes and newlines being printed
intermingled.

We could probably have solved this without any locks:
have the template do its complete execution into a temporary buffer,
and then write the buffer (all at once) to stdout, which is atomic.
But then you may have to worry about managing the temporary buffers.
This is a simple solution that seems to work well.