split off cluster partitioner to separate package
the cluster package is largely the "guts" of the clustering system of
metrictank, which depends on logging and stats packages, memberlist,
etc. Those packages also come with their own dependencies.
The partitioner stuff however is more self contained, and it is
also more commonly imported by other software:
Both tsdb-gw and carbon-relay use the partitioner to send data to kafka,
but use none of the other cluster stuff, so by splitting off the
partitioner we can avoid that they have to pull in a whole bunch of
packages they don't need.