rework archives & consolidation
1) use storage-schemas.conf and storage-aggregation.conf
to configure numchunks, chunkspans, ttls and aggregations.
For carbon also to find out the raw interval.
For other input plugins we only honor interval specified in the data

Note:
- persist message now need to mention which metric names, in case
metrics need to be created on the other end.

2) support configurable retention schemes based on patterns, like
graphite. adjust alignRequests accordingly

3) support configurable aggregation bands to control which rollups
get created per metric pattern.

4) add back 'last' roll-up aggregation and runtime consolidation

It was removed in https://github.com/raintank/metrictank/pull/142
 / 2be35468ea2c486f3c2fc201343a34f79b36055b
based on a discussion in https://github.com/raintank/strategy/issues/11
on the premise that storing this extra series for all our metrics wasn't
worth it and it's not queryable via consolidateBy which only does sum,
avg, min and max.

However:
1) now aggregations are configurable:
   One only enables specific aggregation bands on an as-needed basis.
2) graphite supports last rollups, so we must too.
3) for certain types of data (e.g. counters) it's
   simply the best (better than all the rest) approach
4) storage level aggregation is not as tied to consolidateBy() as we
   originally made it out to be, so that point doesn't really apply
   anymore.
5) we also restore the 'last' runtime consolidation code,
   even though we will no longer do runtime consolidation in MT for now
   but the upcoming normalization uses the same code.
   As we should use 'last' for normalization if rollups was 'last'.
   (see next commit for more info)