update kafkamdm to set priority based on number of seconds lag.
- instead of just using the number of messages the node is lagging
by, divide the number of messages by the ingest rate to get the
number of seconds of lag.  ie, if a node is ingesting at 10k/s
and the lag is 1k, then the priority is set to 0, as the outstanding
metrics will be processed within 1second.  But if the ingest rate
is 10k, and the lag is 40k then the prio gets set to 4 as it will take
4seconds to process the outstanding metrics at the current rate.