fix metricDef fuzzy formula
due to the multiplication by 100, it was subtracting way too much,
so that metricdefs are less likely to be updated.
However the multiplication is good to turn the float value into a useful
factor, but then we have to divide again to compensate.

With this fix applied, we will update the metricdef more frequently like
we were supposed to.