make consolidateBy stable
the main reason is so that we can deploy this, while still receiving
queries from graphie-metrictank which may specify consolidateBy.
we wouldn't want to trigger a proxy request which would cause a
request loop.

Also technically this is correct anyway, we "support" consolidateBy
but we don't support runtime consolidation yet, so consolidateBy is
basically a no-op for now and that's OK.

also make movingAverage non-stable because it's certainly not.