add new mt-replicator tool
When replicating between clusters in different networks it is not
always easy to push from one kafka to another due to Kafka
requiring that publishers send to the configured "advertised_addr"
which probably isnt internet facing.  This new tool consumes from
a local kafka cluster and publishes data to a remote tsdb-gw server.