support more from/to specifications, honor timezone controls.
by pulling in latest `dur` library which nows needs to be passed in
an explicit timezone. It used to use time.Local which we default
to now as well, but we now allow controlling timezone via configuration
and `tz` argument in render requests, just like graphite.
(note: graphite defaults to Chicago which seems rather silly)

fix #673
fix #681