more offset/lag reporting, fix priority reporting when no consumption
if we're starting from oldest/newest and have not consumed anything yet,
still report lag and current offset, since we can know both accurately.

In particular this fixes a bug where on a kafka cluster where nothing
was being produced, nothing was also obviously being consumed, so no lag
was reported, so priority was reported as 10k, so the node was not able
to become ready.