set more aggressive memberlist parameters for chaos environment
these settings are equivalent to default-local but with a 40s
gossip-to-the-dead-time

there's 2 reasons for this:

* before this, you could tell via the cluster dashboard that the cluster
was barely fast enough (and sometimes not all) to get the cluster in the
desired state, e.g. during the isolation test. now it is more reliable
though sometimes slightly too aggressive (sometimes a node that's
supposed to be healthy is split off for a few seconds) but it's hard
to find a balance. currently, mt4 feels split off for about 10s only
(should be about 30s)
whereas the others see mt4 split off for 25s. clearly this could use
more tweaking

* the gossip-to-the-dead-time allows for the cluster to reconverge to
the full healthy cluster, when the network is healthy again
previously, we didn't wait long enough and the cluster would stay split
forever, after the testIsolate test.