make index pruning also prune the tag index
in order to loop over all defs metric defs we loop over m.DefById. This
is better than looping over the tree and the tag index because it
guarantees that we check every MetricDefinition exactly one time, while
this wouldn't be possible on the tag index because on MetricDefinition
can be referred to by many tags.