Add Delete method to chunk cache
Additionally we now need to keep track of which metric keys are archives
of which raw keys. That is necessary because if we clear metrics from
the chunk cache by raw key we'll also want all the related archives to
be cleared.

Furthermore we add a Reset() method to just completely reset the whole
cache. This is much faster than clearing every single metric in the
cache one-by one, because it basically just reinstantiates the cache.