reduce default kafka-mdm channel-buffer-size
in prod we use 1000.
using too high values will cause the sarama library to consume too much
memory when backfilling, which is not reclaimed, leaving a permament
increased memory usage footprint.
see
https://snapshot.raintank.io/dashboard/snapshot/lu1Aw1kPcbx4fk8qdtgZPQ61yyeTwbBk?orgId=2

Even though we plan to switch to the confluent kafka library imminently
(https://github.com/grafana/metrictank/pull/879)
it's worthwhile to set this straight, for the record.