make offset=oldest default in some of the kafka docker environments
for these environments it makes sense, as it allows to test backfill
speed.
For other environments (and the defaults) it does not, because
it is too vulnerable to segments getting GC'd by kafka and MT failing
to consume. we don't want that to be the default behavior, or in
automated stacks like docker-chaos.