Add cache overhead accounting
Prior to this change we were only counting the number of bytes
that the actual data occupied. However, there is a lot of overhead
in our cache system that also needs to be tracked. We were seeing
large differences in the total cache size used and and live heap
allocations.

These changes attempt to track most of the accounting overhead.
The numbers used to track overhead are estimates based on
data type sizes.

 * Added stats for Flat Acccounting, the LRU, and Chunks
 * Added benchmark for LRU allocations

Resolves: #1086
See also: #963