handle write errors consistently
The Bigtable client will retry writes on transient errors. For any
write that does not have a context.WithTimeout, we can just trust that
the write will continue to be retried until it succeeds.  There is no
need for us to re-attempt the write if it fails.  If there is a
context.WithTimeout then we do retry on failure.