corruption remediation part 2: support single-point chunks
by using an interval hint, we know what kind of delta we need
to end up with a valid timestamp.
For this to be effective this requires an interval:

* >1
* that is not a divisor of 16384.

but for chunks that are so long to be affected (e.g. >4h),
that is always the case.

Note: this can be cleaned up when we have
https://github.com/raintank/schema/pull/22