make chunk.Series explicit
the go behavior of unnamed fields and automatically making
method and field lookups work on them is nice, but ambiguous.
It seems better to be very explicit. For one thing, it allows us
to easily find all the places where certain methods and fields
of Series are being called, which helps with making sure we are
race free.

Perhaps later this can be refactored and tsz can be merged into the
chunk package but that's a worry for later.