add method to prepare filters
the new expression types have a getFilter() method which returns a
simple function to pass a metric definition to, the filter then decides
whether this metric definition should be part of the result set or not.
to evaluate a result set we first build a set of filters to then pass
each of those metric definitions to.