disable the findCache for 1min when a flood of new series is received.
When new series are recieved we need to invlidate any findCache
patterns that match the series name. If there are lots of new series
being created, this can have a negative performance impact.  This change
causes the cache to be disabled for 1minute when the rate of new series
is higher than what we can process.