change ChunkWriteRequests to make them marshalable
this changes the chunk write requests so the logic in the stores is less
coupled with the mdata package. it uses a callback function that gets
assigned by the instantiator of the chunk write request to decouple
them.
it also makes the chunk write requests marshalable so we can transfer
them over the wire. this is useful for the whisper importer utility,
because then it can simply transmit the chunk write requests from the
reader to the writer.