change wire format of whisper importer to cwrs
when transferring the data between the whisper-importer-reader and
-writer we now use the same chunk write request type that metrictank
uses internally when submitting writes to its stores. this makes it
easier to add other store types to the importer-writer if we add more to
metrictank in the future.
futhermore, it modifies the chunk write requests so they include a
callback to be called when a chunk has been written. this helps to
decouple the mdata package from the store package by letting the
instantiator of the chunk write request adding any generic callback to
it.