configure the importer-writer the same way as MT
this makes the configuration of the importer-writer consistent with the
way how MT and other MT-tools are configured. this makes sense to do now
because with the added support for the bigtable store and index the
configuration would get too complex if we'd try to come up with a way to
configure all of that via cli arguments, so its easier and more consistent
to just do it in the same way like everything else.