Remove ConfigSetup() call during the Prometheus input plugin Start() method
  During the prometheus plugin initialization code uses the global
variable partitionID pointer instead of a copy of the value.
It means that when the ConfigSetup() is executed the plugin's configuration
structure values are set back to their default i.e. partitionID=0 here.

With this fix the partitionID selection works for the first time.

NB: scylladb was able to handle 15M metrics on a single partition