optionally reject invalid tags
adds new configuration option which decides whether ingested metrics
with invalid tags should be rejected or not. even if rejection is
disabled a new counter will be increased every time when an invalid
metric has been received. this should make it easy for users who are
not sure whether they currently send invalid metrics to MT to figure it
out before they enable the rejection