fix slow ingest problems
This PR adds code to the vendored version of raintank/schema which has
already been merged into raintank/schema from this PR:
https://github.com/raintank/schema/pull/33. We can't yet merge master
from raintank/schema due to other required changes. The main point of
the PR is to add and use the Clone() method in raintank/schema/metric.go
in order to assign the nameWithTags values to a cloned MetricDefinition.

We also change the CloneArchive() method in idx/memory.go to use the
new Clone() method. This should reduce allocations caused by CloneArchive()
calling MetricDefinition.NameWithTags() which was causing problems,
especially during startup.